Move coordinate-specific link helpers to plugin
