+++ glue/plugins/coordinate_helpers/__init__.py
@@ -1,0 +1,5 @@
+try:
+    from .link_helpers import *
+except ImportError:
+    import warnings
+    warnings.warn("Could not import coordinate_helpers plugin, since Astropy>=0.4 is required")
